requirements.txt

# === Core Dependencies (Roop Framework) ===
numpy==1.24.3
opencv-python==4.8.0.74
onnx==1.14.0
insightface==0.7.3
psutil==5.9.5
pillow==10.0.0

# ONNX Runtime (Platform Specific)
onnxruntime==1.15.1; python_version != '3.9' and sys_platform == 'darwin' and platform_machine != 'arm64'
onnxruntime-coreml==1.13.1; python_version == '3.9' and sys_platform == 'darwin' and platform_machine != 'arm64'
onnxruntime-silicon==1.13.1; sys_platform == 'darwin' and platform_machine == 'arm64'
onnxruntime-gpu==1.15.1; sys_platform != 'darwin'

# Additional Tools
tensorflow==2.13.0
opennsfw2==0.10.2
protobuf==4.23.4
tqdm==4.65.0
gfpgan==1.3.8

# GUI (Optional)
tk==0.1.0
customtkinter==5.2.0
tkinterdnd2==0.3.0; sys_platform != 'darwin' and platform_machine != 'arm64'
tkinterdnd2-universal==1.7.3; sys_platform == 'darwin' and platform_machine == 'arm64'

# === Deepfake Detection System ===
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0

# Computer Vision
mediapipe>=0.10.0

# Data Science
scikit-learn>=1.3.0

# Optimization & Visualization
optuna>=3.0.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
